mental health chatbot/backend/requirements.txt
flask
flask-cors
google-generativeai
langchain-community
python-dotenv
sentence-transformers
torch
chromadb
ipykernel
jupyterlab
notebook
pypdf
langchain_chroma
tiktoken